requirements.txt
numpy==1.16.1
pika==0.13.0
mysql-connector==2.2.9
scipy==1.2.1
protobuf==3.7.1